Only define LaunchedInstalledExe if runAfter is true
The CustomAction LaunchedInstalledExe refers to a FileKey for a
non-existent file, mainExecutableFile. Since it's only necessary
for supporting --run-after, this commit stops it being defined
unless runAfter is truthy.

This is not a fix for the underlying issue! However, it isolates
the problem to users of --run-after, which has other bugs that make
it hard to test, so it's a good interim fix.